Revert "feat: add missing integrations endpoints (#668)"
This reverts commit 9ec7cab3b407dcbd500ae90302ca3b0b9355113f.
